requirements/base.txt
# requirements/base.txt
Django==5.0.1
djangorestframework==3.14.0
django-environ==0.11.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
drf-spectacular==0.27.0
Pillow==10.2.0
python-jose==3.3.0
python-decouple==3.8
celery==5.3.6
redis==5.0.1
django-filter==23.5
djangorestframework-simplejwt==5.3.0
django-cors-headers
